moved the helper funcs for creating cards into own file so the api is easier to read
